+++ src/domain_event_world.h
@@ -8,6 +8,7 @@ enum domain_event_status domain_event_world_register_resolvers(struct domain_eve
 void domain_event_world_shutdown(void);
 void domain_event_world_forget_character(struct char_data *ch);
 void domain_event_world_forget_object(struct obj_data *obj);
+struct char_data *domain_event_world_resolve_character(struct domain_entity_handle handle);
 struct domain_entity_handle domain_event_room_handle(room_rnum room);
 struct domain_entity_handle domain_event_character_handle(struct char_data *ch);
 struct domain_entity_handle domain_event_object_handle(struct obj_data *obj);